cant add duplicate name for checkpoints now

+++ app/controllers/checkpoints_controller.rb
@@ -29,34 +29,37 @@ class CheckpointsController < ApplicationController
     #creates the new checkpoint
     @checkpoint = Checkpoint.new(checkpoint_params)
     @checkpoint.events_id = session[:current_event_id]
-    @checkpoint.save 
+    if @checkpoint.save 
 
-    #gets the selected routes
-    @route_ids = params[:selected_routes]
-    session[:linker_route_ids] = @route_ids
-    session[:linker_route_ids_index] = 0
-    session[:linker_check_id] = @checkpoint.id
-  
-    #adds each to route with the checkpoint to the linker table
-    if @route_ids
-      @route_ids.each do |id|
-        
-        @linker = RoutesAndCheckpointsLinker.new
-        @linker.checkpoint_id = @checkpoint.id
-        @linker.distance_from_start = 0
-        @linker.route_id = id
-        @linker.save
-        #spreadsheet = Spreadsheet.new
-        #spreadsheet.add_checkpoint((Route.where(id: id).first), @checkpoint.name)
-      end
-      #finds the next linker and redirects to it
-      @route_id = session[:linker_route_ids].at(session[:linker_route_ids_index])
-      @route = RoutesAndCheckpointsLinker.where(checkpoint_id: session[:linker_check_id], route_id: @route_id)
-      @route.each do |route|
-        redirect_to edit_routes_and_checkpoints_linker_path(route)
+      #gets the selected routes
+      @route_ids = params[:selected_routes]
+      session[:linker_route_ids] = @route_ids
+      session[:linker_route_ids_index] = 0
+      session[:linker_check_id] = @checkpoint.id
+    
+      #adds each to route with the checkpoint to the linker table
+      if @route_ids
+        @route_ids.each do |id|
+          
+          @linker = RoutesAndCheckpointsLinker.new
+          @linker.checkpoint_id = @checkpoint.id
+          @linker.distance_from_start = 0
+          @linker.route_id = id
+          @linker.save
+          #spreadsheet = Spreadsheet.new
+          #spreadsheet.add_checkpoint((Route.where(id: id).first), @checkpoint.name)
+        end
+        #finds the next linker and redirects to it
+        @route_id = session[:linker_route_ids].at(session[:linker_route_ids_index])
+        @route = RoutesAndCheckpointsLinker.where(checkpoint_id: session[:linker_check_id], route_id: @route_id)
+        @route.each do |route|
+          redirect_to edit_routes_and_checkpoints_linker_path(route)
+        end
+      else
+        redirect_to checkpoints_path
       end
     else
-      redirect_to checkpoints_path
+      render :new
     end
   end
 
@@ -64,45 +67,48 @@ class CheckpointsController < ApplicationController
   def update
     spreadsheet = Spreadsheet.new
     old_checkpoint_name = @checkpoint.name.dup
-    @checkpoint.update(checkpoint_params)
+    if @checkpoint.update(checkpoint_params)
 
-    #get the selected routes
-    @route_ids = params[:selected_routes]
-    @route_ids.each do |id|
-      spreadsheet.update_checkpoint_name((Route.where(id: id).first), old_checkpoint_name, @checkpoint.name)
-    end
-    session[:linker_route_ids] = @route_ids
-    session[:linker_route_ids_index] = 0
-    session[:linker_check_id] = @checkpoint.id
-
-    #check if the routes are already in the linker with the checkpoint
-    @routes = Route.where(events_id: @checkpoint.events_id)
-    @routes.each do |route|
-      if RoutesAndCheckpointsLinker.exists?(checkpoint_id: session[:linker_check_id], route_id: route.id) && (!@route_ids || !(session[:linker_route_ids].include? (route.id).to_s))
-        #delete from linker table
-        @linker = RoutesAndCheckpointsLinker.where(checkpoint_id: session[:linker_check_id], route_id: route.id)
-        @linker.each do |linker|
-          #TODO need to change this to delete checkpoint in spreadsheet
-          #spreadsheet.update_checkpoint_name((Route.where()), old_checkpoint_name, @checkpoint.name)
-          linker.destroy
-        end
-      elsif !RoutesAndCheckpointsLinker.exists?(checkpoint_id: session[:linker_check_id], route_id: route.id) && (@route_ids && (session[:linker_route_ids].include? (route.id).to_s)) 
-        #add to linker table
-        @linker = RoutesAndCheckpointsLinker.new
-        @linker.checkpoint_id = @checkpoint.id
-        @linker.route_id = route.id
-        @linker.distance_from_start = 0
-        @linker.save
-        #spreadsheet.update_checkpoint_name((Route.where()), old_checkpoint_name, @checkpoint.name)
+      #get the selected routes
+      @route_ids = params[:selected_routes]
+      @route_ids.each do |id|
+        spreadsheet.update_checkpoint_name((Route.where(id: id).first), old_checkpoint_name, @checkpoint.name)
       end
-    end 
-    #finds the next linker and redirects to it or goes to index
-    if @route_ids
-      @route_id = session[:linker_route_ids].at(session[:linker_route_ids_index])
-      @route = RoutesAndCheckpointsLinker.where(checkpoint_id: session[:linker_check_id], route_id: @route_id).first
-      redirect_to edit_routes_and_checkpoints_linker_path(@route)
+      session[:linker_route_ids] = @route_ids
+      session[:linker_route_ids_index] = 0
+      session[:linker_check_id] = @checkpoint.id
+
+      #check if the routes are already in the linker with the checkpoint
+      @routes = Route.where(events_id: @checkpoint.events_id)
+      @routes.each do |route|
+        if RoutesAndCheckpointsLinker.exists?(checkpoint_id: session[:linker_check_id], route_id: route.id) && (!@route_ids || !(session[:linker_route_ids].include? (route.id).to_s))
+          #delete from linker table
+          @linker = RoutesAndCheckpointsLinker.where(checkpoint_id: session[:linker_check_id], route_id: route.id)
+          @linker.each do |linker|
+            #TODO need to change this to delete checkpoint in spreadsheet
+            #spreadsheet.update_checkpoint_name((Route.where()), old_checkpoint_name, @checkpoint.name)
+            linker.destroy
+          end
+        elsif !RoutesAndCheckpointsLinker.exists?(checkpoint_id: session[:linker_check_id], route_id: route.id) && (@route_ids && (session[:linker_route_ids].include? (route.id).to_s)) 
+          #add to linker table
+          @linker = RoutesAndCheckpointsLinker.new
+          @linker.checkpoint_id = @checkpoint.id
+          @linker.route_id = route.id
+          @linker.distance_from_start = 0
+          @linker.save
+          #spreadsheet.update_checkpoint_name((Route.where()), old_checkpoint_name, @checkpoint.name)
+        end
+      end 
+      #finds the next linker and redirects to it or goes to index
+      if @route_ids
+        @route_id = session[:linker_route_ids].at(session[:linker_route_ids_index])
+        @route = RoutesAndCheckpointsLinker.where(checkpoint_id: session[:linker_check_id], route_id: @route_id).first
+        redirect_to edit_routes_and_checkpoints_linker_path(@route)
+      else
+        redirect_to checkpoints_path
+      end
     else
-      redirect_to checkpoints_path
+      render :edit
     end
   end
